test(design-system): extract CSS helper utilities for token tests
- Extract `extractCSSVariables` function to shared `css-helpers.ts` utility module
- Remove duplicate CSS variable extraction logic from colors.test.ts, effects.test.ts, spacing.test.ts, and typography.test.ts
- Fix trailing whitespace and formatting inconsistencies across all token test files
- Improve test maintainability by centralizing CSS parsing logic for reuse across token validation tests
