+++ src/styles/tokens/spacing.test.ts
@@ -1,6 +1,6 @@
 /**
  * Property-Based Tests for Spacing System Tokens
- * 
+ *
  * Feature: design-guideline
  * Tests validate correctness properties from the design document.
  */
@@ -9,43 +9,11 @@ import * as fc from 'fast-check';
 import * as fs from 'fs';
 import * as path from 'path';
 import { describe, expect, it } from 'vitest';
+import { extractCSSVariables } from '../../test-utils/css-helpers';
 
 // Read and parse the spacing.css file
 const spacingPath = path.join(__dirname, 'spacing.css');
 const spacingCSS = fs.readFileSync(spacingPath, 'utf-8');
-
-/**
- * Extract CSS custom property values from the CSS content
- * Uses brace-matching to correctly handle nested CSS blocks
- */
-function extractCSSVariables(css: string, selector: string = ':root'): Map<string, string> {
-  const variables = new Map<string, string>();
-  const escapedSelector = selector.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-  const selectorRegex = new RegExp(`${escapedSelector}\\s*\\{`, 'g');
-  
-  let match;
-  while ((match = selectorRegex.exec(css)) !== null) {
-    // Find matching closing brace with proper nesting
-    let depth = 1;
-    const start = match.index + match[0].length;
-    let i = start;
-    
-    while (i < css.length && depth > 0) {
-      if (css[i] === '{') depth++;
-      else if (css[i] === '}') depth--;
-      i++;
-    }
-    
-    const block = css.slice(start, i - 1);
-    const varRegex = /--([\w-]+):\s*([^;]+);/g;
-    let varMatch;
-    while ((varMatch = varRegex.exec(block)) !== null) {
-      variables.set(`--${varMatch[1]}`, varMatch[2].trim());
-    }
-  }
-  
-  return variables;
-}
 
 /**
  * Parse a pixel value to a number
@@ -56,7 +24,7 @@ function parsePixelValue(value: string): number | null {
   if (value.startsWith('var(')) {
     return null;
   }
-  
+
   const match = value.match(/^(\d+)px$/);
   if (match) {
     return parseInt(match[1], 10);
@@ -81,10 +49,10 @@ for (const [key, value] of rootVariables) {
 describe('Spacing System Properties', () => {
   /**
    * Feature: design-guideline, Property 1: Spacing Scale Base-4 Invariant
-   * 
-   * *For any* spacing value defined in the spacing system, that value SHALL 
+   *
+   * *For any* spacing value defined in the spacing system, that value SHALL
    * be a multiple of 4 pixels.
-   * 
+   *
    * **Validates: Requirements 3.1**
    */
   describe('Property 1: Spacing Scale Base-4 Invariant', () => {
@@ -114,7 +82,7 @@ describe('Spacing System Properties', () => {
     it('should have the expected spacing scale values', () => {
       const expectedScale = [0, 4, 8, 12, 16, 24, 32, 48, 64, 96];
       const actualValues = spacingValues.map((s) => s.value).sort((a, b) => a - b);
-      
+
       expect(actualValues).toEqual(expectedScale);
     });
   });
